Rename jest_matchers_object.js -> jestMatchersObject.js
- Rename file, imports
- No test found
