+++ packages/expect/src/extractExpectedAssertionsErrors.js
@@ -14,7 +14,7 @@ import {
   pluralize,
 } from 'jest-matcher-utils';
 
-import {getState, setState} from './jest_matchers_object';
+import {getState, setState} from './jestMatchersObject';
 
 const resetAssertionsLocalState = () => {
   setState({
